Fix filter by type function finished
